+++ ChangeLog.md
@@ -22,6 +22,11 @@
 - Fix the generation of the root `__init__` files for packages with only models #1195
 - Add pylint and mypy support for `--version-tolerant` generations with `--models-mode=msrest` #1202
 
+**Breaking Changes in Version Tolerant Generation**
+
+- Change the models filename from `_models_py3.py` to `_models.py` #1204
+- Change the enums filename to `_enums.py` #1204
+
 ### 2022-03-08 - 5.14.0
 
 | Library                                                                 | Min Version |